AUTH [#234] refactor(auth): update component imports and restructure layout for account features
