--- containers/frontend/web/src/app/[locale]/(guest)/(auth)/resend-verification/page.tsx
@@ -1,5 +1,0 @@
-import { ResendVerificationFeature } from '@/features/auth/resend-verification';
-
-export default function ResendVerificationPage() {
-  return <ResendVerificationFeature />;
-}
