+++ containers/frontend/web/src/app/[locale]/(guest)/(auth)/verify-email/page.tsx
@@ -1,4 +1,4 @@
-import { VerifyEmailFeature } from '@/features/auth/verify-email';
+import { VerifyEmailFeature } from '@/features/auth/create-account';
 
 export default function VerifyEmailPage() {
   return <VerifyEmailFeature />;
